Create a "UnableToExportError" exception for known exceptions
